Move SymType out of inspect module
The SymType enum is really fairly generic and not specific to the
inspect set of APIs. As such, it makes more sense to have it be defined
in and exposed by the library itself as opposed to a sub-module.
This change moves it out of the inspect module and into lib.rs.
